add hashmap implementation
