+++ backend/src/algo/index.ts
@@ -1,16 +1,28 @@
-import { GenerateScheduleProps, GenerateScheduleResult } from './types';
-import { getConflictingSubjectIndexes } from './utils';
+type HashmapKey = string | object;
+type HashmapValue = object;
 
-export function generateSchedule({ classes, classrooms }: GenerateScheduleProps): GenerateScheduleResult {
-  const subjects = classes.map(c => c.subjects.map(s => ({ ...s, class: c.name }))).flat();
+interface IHashmap {
+  [key: string]: HashmapValue;
 
-  const conflictingIds = [
-    ...getConflictingSubjectIndexes(subjects, s => s.class),
-    ...getConflictingSubjectIndexes(subjects, s => s.teacher?.name),
-    ...getConflictingSubjectIndexes(subjects, s => s.classroom?.name),
-  ];
+  get(key: HashmapKey): HashmapValue;
+  set(key: HashmapKey, value: HashmapValue): void;
+}
 
-  console.log(conflictingIds);
+function Hashmap(): IHashmap {
+  return {
+    get(key: HashmapKey): HashmapValue {
+      return this[JSON.stringify(key)];
+    },
+    set(key: HashmapKey, value: HashmapValue) {
+      this[JSON.stringify(key)] = value;
+    },
+  };
+}
 
-  return { classes: [], teachers: [] };
+export function generateSchedule(_p: any): any {
+  const hm = Hashmap();
+
+  hm.set([[1, 1, 2], 'a3'], ['matematika', 'mate', 3, 'c']);
+
+  return hm.get([[1, 1, 2], 'a3']);
 }
